api/app.py
from fastapi import FastAPI, UploadFile, File
import uvicorn
import numpy as np
from PIL import Image
from io import BytesIO
import tensorflow as tf
from fastapi.middleware.cors import CORSMiddleware



app = FastAPI()

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # or ["http://localhost:3000"] for more security
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

MODEL_PATH = r"models\model 1.keras"
MODEL = tf.keras.models.load_model(MODEL_PATH)
CLASSNAMES= ["Early Blight", "Healthy", "Late Blight"]
IMAGE_SIZE = 256

@app.get("/hello")
async def hello():
    return {"message": "Hello, World!"}

def read_file_as_image(data) -> np.ndarray:
    """Convert file data to an image array."""
    image = Image.open(BytesIO(data)).convert("RGB")
    image = image.resize((IMAGE_SIZE, IMAGE_SIZE))  # Resize to 256x256
    image = np.array(image) / 255.0  # Normalize to [0,1]
    return image

@app.post("/predict")
async def predict(file: UploadFile = File(...)):
    image = read_file_as_image(await file.read())
    
    img_batch = np.expand_dims(image, 0)
    
    predictions = MODEL.predict(img_batch)

    predicted_class = CLASSNAMES[np.argmax(predictions[0])]
    confidence = np.max(predictions[0])
    return {
        'class': predicted_class,
        'confidence': float(confidence)
    }

if __name__ == "__main__":
    uvicorn.run(app, host="localhost", port=8000)
