requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
tensorflow==2.19.0
keras==3.10.0
pillow==11.3.0
numpy==2.1.3
python-multipart==0.0.20
